versión 2
actualización del curriculum vitae

+++ app/Http/Controllers/HomeController.php
@@ -37,21 +37,69 @@ class HomeController extends Controller
         $password_update = ($request->password !=$user->password)? Hash::make($request->password) : $user->password;
         $header_image_path_update = $user->header_image_path;
         $my_perfil_update = $user->my_perfil;
+        $slogan_es = $user->slogan_es;
+        $slogan_en = $user->slogan_en;
+        $logo = $user->logo;
+        $avatar = $user->avatar;
         //agregar foto para el header de la pagina
         if(isset($request->header_image_path)&&!empty($request->header_image_path)){
-            $path = explode("/", $user->header_image_path);
-            Storage::delete("public/{$path[2]}");
+            if( $user->header_image_path)
+            {
+                $path = explode("/", $user->header_image_path);
+                Storage::delete("public/{$path[2]}");
+            }
             $image = $request->file('header_image_path')->store('public');
             $header_image_path_update = Storage::url($image);
         }
         //Agregar una foto de perfil
         if(isset($request->my_perfil)&&!empty($request->my_perfil)){
-            $path = explode("/", $user->my_perfil);
-            Storage::delete("public/{$path[2]}");
+            if( $user->my_perfil){
+                $path = explode("/", $user->my_perfil);
+                Storage::delete("public/{$path[2]}");
+            }
             $image = $request->file('my_perfil')->store('public');
             $my_perfil_update = Storage::url($image);
         }
-
+        //Agregar una foto de eslogan en
+        if(isset($request->slogan_en)&&!empty($request->slogan_en)){
+            if($user->slogan_en){
+                $path = explode("/", $user->slogan_en);
+                Storage::delete("public/{$path[2]}");
+            }
+            $image = $request->file('slogan_en')->store('public');
+            $slogan_en = Storage::url($image);
+        }
+        //Agregar una foto de eslogan es
+        if(isset($request->slogan_es)&&!empty($request->slogan_es)){
+            if($user->slogan_es)
+            {
+                $path = explode("/", $user->slogan_es);
+                Storage::delete("public/{$path[2]}");
+            }
+            $image = $request->file('slogan_es')->store('public');
+            $slogan_es = Storage::url($image);
+        }
+        //Agregar una foto de logo
+        if(isset($request->logo)&&!empty($request->logo)){
+            if($user->logo)
+            {
+                $path = explode("/", $user->logo);
+                Storage::delete("public/{$path[2]}");
+            }
+            $image = $request->file('logo')->store('public');
+            $logo = Storage::url($image);
+        }
+        //Agregar una foto del avatar
+        if(isset($request->avatar)&&!empty($request->avatar)){
+            if($user->avatar)
+            {
+                $path = explode("/", $user->avatar);
+                Storage::delete("public/{$path[2]}");
+            }
+            $image = $request->file('avatar')->store('public');
+            $avatar = Storage::url($image);
+        }
+//'slogan_en','slogan_es','logo'
         if(User::where('id',Auth::id())->update(
             [
                 'name'=>$request->name,
@@ -66,7 +114,13 @@ class HomeController extends Controller
                 'country_es'=>$request->country_es,
                 'country_en'=>$request->country_en,
                 'header_image_path'=>$header_image_path_update,
-                'my_perfil'=>$my_perfil_update
+                'my_perfil'=>$my_perfil_update,
+                'slogan_en'=>$slogan_en,
+                'slogan_es'=>$slogan_es,
+                'logo'=>$logo,
+                'avatar'=>$avatar,
+                'header_text_en'=>$request->header_text_en,
+                'header_text_es'=>$request->header_text_es,
             ]
         )){
             return redirect()->route('home')->with('success','Actualizado correctamente');
